submit hashed task list also modularized
